1) 코드 가독성 향상 작업 convert.py 파일 분리
